requirements.txt
boto3==1.26.24
pendulum==2.1.2
apache-airflow-providers-amazon==7.3.0